Rename module to tango_ros_common for consistency
