+++ tango_ros_common/tango_ros_common/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
 
-    package="eu.intermodalics.tangoroscommon">
+    package="eu.intermodalics.tango_ros_common">
 
     <application android:allowBackup="true" android:label="@string/app_name"
         android:supportsRtl="true">
