+++ tango_ros_common/tango_ros_common/src/main/java/eu/intermodalics/tango_ros_common/Logger.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package eu.intermodalics.tangoroscommon;
+package eu.intermodalics.tango_ros_common;
 
 import android.app.Activity;
 import android.os.Environment;
